Documentation and linting
