+++ app/src/main/java/de/tum/in/tumcampusapp/utils/camera/ImageUtils.kt
@@ -91,7 +91,7 @@ object ImageUtils {
         } catch (e: IOException) {
             Utils.log(e)
         }
-        return false;
+        return false
     }
 
     /**
